aesthetic updates to ourteam / finished contactus yayyy
