+++ src/routes/contactus.css
@@ -1,34 +1,21 @@
-.ContactUsWrapper {
-    background-color: #000;
-    padding: 3%;
-}
 
 .ContactUs {
     display: flex;
     flex-direction: column;
-    padding-top: 80px;
     align-items: flex-start;
     justify-content: space-around;
     font-size: 3rem;
-    color: white;
+    color: black;
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    padding-top: 5rem;
 }
 
 .ContactUsTitle {
     font-size: 120px;
-}
-
-.ContactUsInput {
-    padding: 5%;
-}
-
-.ContactUsSendButton {
-    background: black;
-    color: white;
-}
-
-.ContactUsSendButton:hover {
-    background: white;
-    color: black
+    margin-bottom: 2rem;
 }
 
 .ContactUsInfo {
