+++ src/routes/ourteam.css
@@ -9,7 +9,7 @@
 
 .team-message-one {
     background-color: rgb(179, 174, 245);
-    font-size: 60px;
+    font-size: 42px;
     font-weight: 700; 
     color: black;
     font-style: italic;
