+++ src/routes/ourteam.js
@@ -14,7 +14,7 @@ export default function OurTeam() {
                 <mark className='team-message-one'>WITH OUR APPROACH, WE NURTURE CHILDREN'S CURIOSITY AND TURN THEIR POTENTIAL INTO SKILLS THAT HELP THEM GROW CONFIDENTLY, CREATING A STRONG FOUNDATION FOR THEIR FUTURE.</mark>
             </p>
             <TheTeam/>
-            <p style={{padding: '3%', textAlign: 'center', fontSize: '50px', fontWeight: '700', marginTop: '70px', marginBottom: '110px'}}>OUR TEAM IS MORE THAN JUST CARING FACES AND EDUCATORS. WE PROVIDE A SAFE SPACE FOR CHILDREN TO EXPLORE AND DISCOVER THE WORLD AROUND THEM.</p>
+            <p style={{padding: '3%', textAlign: 'center', fontSize: '42px', fontWeight: '700', marginTop: '70px', marginBottom: '110px'}}>OUR TEAM IS MORE THAN JUST CARING FACES AND EDUCATORS. WE PROVIDE A SAFE SPACE FOR CHILDREN TO EXPLORE AND DISCOVER THE WORLD AROUND THEM.</p>
             <Footer/>
         </div>
     );
